+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/abstracts/WithChat.kt
@@ -1,0 +1,10 @@
+package dev.inmo.tgbotapi.abstracts
+
+import dev.inmo.tgbotapi.types.chat.Chat
+
+/**
+ * All inheritors of this interface have [chat] field and related to this [chat]
+ */
+interface WithChat {
+    val chat: Chat
+}