+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # TelegramBotAPI changelog
 
+## 7.0.1
+
+* `Core`:
+    * New interface `WithChat` which contains `chat` field
+        * `Message` now inherits `WithChat`
+        * `ChatMemberUpdated` now inherits `WithChat`
+* `Utils`:
+    * Improvements in `same`-notations
+
 ## 7.0.0
 
 This update contains support of [Telegram Bot API 6.6](https://core.telegram.org/bots/api-changelog#march-9-2023)